Explicitly cast size_t to unsigned int, to fix a warning about that.

+++ Spinblocks/src/Systems/EliminateSystem.cpp
@@ -31,7 +31,7 @@ namespace Systems
 
 				if (moveable.GetCurrentCoordinate().Get().y > lowestRow)
 				{
-					moveable.SetDesiredCoordinate(GetCoordinateOfEntity(registry, MoveBlockInDirection(registry, entity, moveDirection_t::SOUTH, rows.size(), true)));
+					moveable.SetDesiredCoordinate(GetCoordinateOfEntity(registry, MoveBlockInDirection(registry, entity, moveDirection_t::SOUTH, static_cast<unsigned int>(rows.size()), true)));
 					moveable.SetMovementState(Components::movementStates_t::HARD_DROP);
 				}
 			}
